Выполнение задания 5.15
