+++ index.html
@@ -232,5 +232,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 <script src="main.js"></script>
+<script type="module" src="mother.js"></script>
+<script type="module" src="сain.js"></script>
 </body>
 </html>
