+++ js/main.js
@@ -1,41 +1,12 @@
-//task1
+import {generatePhoto} from './data.js';
+import {getNumber} from './utility.js';
 
-function returnNumberFrom(a, b){
-  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
-  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-}
-
-//task2
-//eslint-disable-next-line
-function checkStringLength(string, maxLenght){
-  return string.length <= maxLenght;
-}
-
-/*
-module 4-task1
-*/
-const number = getNumber(25);
-const array = new Array(number);
+const specifiedNumber = getNumber(25);
+const array = new Array(specifiedNumber);
 
 for (let i = 1; i <= array.length; i++){
   const element = generatePhoto(i,i);
   array[i-1] = element;
 }
 
-function generatePhoto(id, url){
-  const descriptions = ['Поползень','Гастеруптиида','Лягушкорот','Квакша','Игуанодонт'];
-  const urlString = `photos/${url}.jpg`;
-  const photo = {
-    id: id,
-    url: urlString,
-    description: descriptions[returnNumberFrom(0,(descriptions.length - 1))],
-    likes: returnNumberFrom(15,200),
-    comments: returnNumberFrom(0,200),
-  };
-  return photo;
-}
-function getNumber(num){
-  return num;
-}
+
