+++ js/utility.js
@@ -1,0 +1,17 @@
+function returnNumberFrom(a, b){
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+//eslint-disable-next-line
+function checkStringLength(string, maxLenght){
+  return string.length <= maxLenght;
+}
+
+function getNumber(num){
+  return num;
+}
+
+export {returnNumberFrom, getNumber};
